Improve null checks and enhance type safety across multiple files
- Refined null checks in `geometry.cpp`, `intersectionchecks.cpp`, and `livingentity.cpp` to ensure better clarity and consistency.
- Updated type casting in `aabbtree.cpp`, `obbtree.cpp`, and `utils.cpp` to use `intptr_t` for safer pointer arithmetic.
- Enhanced matrix handling in `softentity.cpp` and `ropeentity.cpp` for improved readability and maintainability.
- Added missing `minmax` function in `aabbtree.cpp` and `obbtree.cpp` to standardize utility functions across the codebase.
- Improved initialization and assignment operations in `physicalentity.cpp` and `matrixnm.h` for better performance and clarity.

+++ CryPhysics/geometry.cpp
@@ -586,9 +586,9 @@ int CPrimitive::Intersect(IGeometry *_pCollider, geom_world_data *pdata1,geom_wo
 
 	static geom_world_data defgwd;
 	static intersection_params defip;
-	pdata1 = (geom_world_data*)((intptr_t)pdata1 | -iszero((intptr_t)pdata1) & (intptr_t)&defgwd);
-	pdata2 = (geom_world_data*)((intptr_t)pdata2 | -iszero((intptr_t)pdata2) & (intptr_t)&defgwd);
-	pparams = (intersection_params*)((intptr_t)pparams | -iszero((intptr_t)pparams) & (intptr_t)&defip);
+	pdata1 = (geom_world_data*)((intptr_t)pdata1 | -iszero((int)(intptr_t)pdata1) & (intptr_t)&defgwd);
+	pdata2 = (geom_world_data*)((intptr_t)pdata2 | -iszero((int)(intptr_t)pdata2) & (intptr_t)&defgwd);
+	pparams = (intersection_params*)((intptr_t)pparams | -iszero((int)(intptr_t)pparams) & (intptr_t)&defip);
 	if (!pparams->bKeepPrevContacts)
 		g_nAreas = g_nAreaPt = g_nTotContacts = g_BrdPtBufPos = 0;
 	if (g_nTotContacts>=sizeof(g_Contacts)/sizeof(g_Contacts[0]))
